+++ R/GBIFtableCleanup.R
@@ -26,9 +26,9 @@ GBIFtableCleanup <- function(GBIFtable) {
     GBIFtable["day"] <- NA
     GBIFtable["month"] <- NA
     GBIFtable["year"] <- NA
-    GBIFtable["Dataset"] <- NA
-    GBIFtable["DatasetKey"] <- NA
-    GBIFtable["DataService"] <- NA
+    GBIFtable["datasetName"] <- NA
+    GBIFtable["datasetKey"] <- NA
+    GBIFtable["dataService"] <- NA
     GBIFtable <- as.data.frame(as.list(GBIFtable))
     return(GBIFtable)
   } else if (nrow(GBIFtable) == 0) {
@@ -40,9 +40,9 @@ GBIFtableCleanup <- function(GBIFtable) {
     GBIFtable["day"] <- NA
     GBIFtable["month"] <- NA
     GBIFtable["year"] <- NA
-    GBIFtable["Dataset"] <- NA
-    GBIFtable["DatasetKey"] <- NA
-    GBIFtable["DataService"] <- NA
+    GBIFtable["datasetName"] <- NA
+    GBIFtable["datasetKey"] <- NA
+    GBIFtable["dataService"] <- NA
     GBIFtable <- as.data.frame(as.list(GBIFtable))
     return(GBIFtable)
   } else {
@@ -55,9 +55,9 @@ GBIFtableCleanup <- function(GBIFtable) {
       GBIFtable["day"] <- as.integer(unlist(GBIFtable["day"]))
       GBIFtable["month"] <- as.integer(unlist(GBIFtable["month"]))
       GBIFtable["year"] <- as.integer(unlist(GBIFtable["year"]))
-      GBIFtable["Dataset"] <- as.factor(unlist(GBIFtable["Dataset"]))
-      GBIFtable["DatasetKey"] <- as.factor(unlist(GBIFtable["DatasetKey"]))
-      GBIFtable["DataService"] <- as.factor(unlist(GBIFtable["DataService"]))
+      GBIFtable["datasetName"] <- as.factor(unlist(GBIFtable["datasetName"]))
+      GBIFtable["datasetKey"] <- as.factor(unlist(GBIFtable["datasetKey"]))
+      GBIFtable["dataService"] <- as.factor(unlist(GBIFtable["dataService"]))
       return(GBIFtable)
     } else {
       GBIFtable <- NULL
@@ -68,9 +68,9 @@ GBIFtableCleanup <- function(GBIFtable) {
       GBIFtable["day"] <- NA
       GBIFtable["month"] <- NA
       GBIFtable["year"] <- NA
-      GBIFtable["Dataset"] <- NA
-      GBIFtable["DatasetKey"] <- NA
-      GBIFtable["DataService"] <- NA
+      GBIFtable["datasetName"] <- NA
+      GBIFtable["datasetKey"] <- NA
+      GBIFtable["dataService"] <- NA
       GBIFtable <- as.data.frame(as.list(GBIFtable))
       return(GBIFtable)
     }