+++ tests/testthat/test-map.R
@@ -2,29 +2,21 @@ context("Testing mapping function")
 
 library(occCite)
 
+data("myOccCiteObject")
+
 test_that("inputs to map are as expected", {
-  data("myOccCiteObject")
-  testResults <- myOccCiteObject@occResults
-  resNames <- names(testResults)
-  sp.name <- names(myOccCiteObject@occResults)[[1]]
   x <- myOccCiteObject@occResults[[1]]
-  tabResults <- tabulate.occResults(x = x, sp.name = sp.name)
+  sp.name <- names(myOccCiteObject@occResults)[[1]]
+  tabResults <- occCite:::tabulate.occResults(x = x, sp.name = sp.name)
 
   expect_true(class(myOccCiteObject) == "occCiteData")
   expect_true(names(myOccCiteObject@occResults) > 0)
-  expect_true(is.character(resNames))
-  expect_true(all(!is.na(stringr::str_extract(
-    string = resNames,
-    pattern = "(\\w+\\s\\w+)"
-  ))))
 
   expect_true(all(c("longitude", "latitude")
-  %in% names(tabulate.occResults(x, sp.name))))
+  %in% names(occCite:::tabulate.occResults(x, sp.name))))
 })
 
 test_that("default occCiteMap settings work", {
-  data("myOccCiteObject")
-
   expect_error(occCiteMap())
 
   test <- occCiteMap(myOccCiteObject)
@@ -32,8 +24,6 @@ test_that("default occCiteMap settings work", {
 })
 
 test_that("occCiteMap works with species specified", {
-  data("myOccCiteObject")
-
   expect_error(occCiteMap(myOccCiteObject, "Protea cynaroides",
                           species_colors = "red", "blue"))
 
@@ -42,7 +32,6 @@ test_that("occCiteMap works with species specified", {
 })
 
 test_that("occCiteMap works with non-awesome markers color specified", {
-  data("myOccCiteObject")
   test <- occCiteMap(myOccCiteObject,
     "Protea cynaroides",
     species_colors = "brown",
@@ -52,7 +41,6 @@ test_that("occCiteMap works with non-awesome markers color specified", {
 })
 
 test_that("occCiteMap works with awesome markers color specified", {
-  data("myOccCiteObject")
   test <- occCiteMap(myOccCiteObject,
     "Protea cynaroides",
     species_colors = "lightred",
@@ -62,7 +50,6 @@ test_that("occCiteMap works with awesome markers color specified", {
 })
 
 test_that("occCiteMap works with map_limit specified", {
-  data("myOccCiteObject")
   test <- occCiteMap(myOccCiteObject,
     "Protea cynaroides",
     species_colors = "lightred",
@@ -73,7 +60,6 @@ test_that("occCiteMap works with map_limit specified", {
 })
 
 test_that("occCiteMap works with cluster set to true", {
-  data("myOccCiteObject")
   test <- occCiteMap(myOccCiteObject,
     "Protea cynaroides",
     species_colors = "lightred",
@@ -82,3 +68,4 @@ test_that("occCiteMap works with cluster set to true", {
   )
   expect_true(all(c("leaflet", "htmlwidget") %in% class(test)))
 })
+