+++ NEWS.md
@@ -1,6 +1,9 @@
 # occCite 0.5.7
 
-* Development version
+* Fixed warning in `occCitation()` when getting GBIF citations
+* Updated date formatting using `format()` instead of `as.character()`
+* Updated test files to incorporate rgbif 3.8.0 output changes, make more efficient
+* In occResults, renamed "Dataset", "DatasetKey", and "DataService" to "datasetName", "datasetKey", "dataService", respectively.
 
 # occCite 0.5.6